Solve issue #95
I have looked into issue #95 on an instance running CiviCRM 5.38.1 with version 1.2 of this extension. In the code section that constructs the HTML for contact references, the array ``$field`` did not contain a member called ``contact_ref_id``. Instead, there was a one-element array called ``contact_ref_links`` whose first member was a string with a ready-made HTML link. I figured that the previous solution probably worked at some point, so I added code that prefers ``contact_ref_links[0]``if it is available (if the data provides a link already, it should probably be used, not constructed manually as the old code did) but the code will fall back to the old one if ``contact_ref_links`` is not provided but ``contact_ref_id`` is. If neither is available, a link cannot be constructed, in which case only ``$field['value]`` is returned (which in my test contained the contact name). I did not look into where the data in ``$field`` came from but I suspect it’s from a source outside of this extension. Maybe with more insight a different solution might seem preferrable but from my point of view this is the best strategy.